+++ networkx/algorithms/traversal/depth_first_search.py
@@ -2,36 +2,34 @@
 ==================
 Depth-first search
 ==================
-
-Basic algorithms for depth-first searching the nodes of a graph.
-
+Basic algorithms for depth-first search and depth-limited search of a graph.
 Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
 by D. Eppstein, July 2004.
+https://en.wikipedia.org/wiki/Depth-limited_search
 """
 import networkx as nx
 from collections import defaultdict
 __author__ = """\n""".join(['Aric Hagberg <aric.hagberg@gmail.com>'])
 __all__ = ['dfs_edges', 'dfs_tree',
            'dfs_predecessors', 'dfs_successors',
-           'dfs_preorder_nodes','dfs_postorder_nodes',
+           'dfs_preorder_nodes', 'dfs_postorder_nodes',
            'dfs_labeled_edges']
 
-def dfs_edges(G, source=None):
-    """Produce edges in a depth-first-search (DFS).
 
+def dfs_edges(G, source=None, depth_limit=None):
+    """Produce edges in a depth-first-search (DFS).
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     edges: generator
        A generator of edges in the depth-first-search.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -43,7 +41,7 @@ def dfs_edges(G, source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
@@ -53,38 +51,42 @@ def dfs_edges(G, source=None):
     else:
         # produce edges for components with source
         nodes = [source]
-    visited=set()
+    visited = set()
+    if depth_limit is None:
+       depth_limit = len(G)
     for start in nodes:
         if start in visited:
             continue
         visited.add(start)
-        stack = [(start,iter(G[start]))]
+        stack = [(start, depth_limit, iter(G[start]))]
         while stack:
-            parent,children = stack[-1]
+            parent, depth_now, children = stack[-1]
+            if not depth_now >= 1:
+                stack.pop()
+                continue
             try:
                 child = next(children)
                 if child not in visited:
-                    yield parent,child
+                    yield parent, child
                     visited.add(child)
-                    stack.append((child,iter(G[child])))
+                    stack.append((child, depth_now-1, iter(G[child])))
             except StopIteration:
                 stack.pop()
 
-def dfs_tree(G, source):
-    """Return oriented tree constructed from a depth-first-search from source.
 
+def dfs_tree(G, source, depth_limit=None):
+    """Return oriented tree constructed from a depth-first-search from source.
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     T : NetworkX DiGraph
        An oriented tree
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -98,25 +100,24 @@ def dfs_tree(G, source):
         T.add_nodes_from(G)
     else:
         T.add_node(source)
-    T.add_edges_from(dfs_edges(G,source))
+    T.add_edges_from(dfs_edges(G, source, depth_limit))
     return T
 
-def dfs_predecessors(G, source=None):
-    """Return dictionary of predecessors in depth-first-search from source.
 
+def dfs_predecessors(G, source=None, depth_limit=None):
+    """Return dictionary of predecessors in depth-first-search from source.
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     pred: dict
        A dictionary with nodes as keys and predecessor nodes as values.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -128,29 +129,28 @@ def dfs_predecessors(G, source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
-    return dict((t,s) for s,t in dfs_edges(G,source=source))
+    return dict((t, s) for s, t in dfs_edges(G, source=source,
+    depth_limit=depth_limit))
 
 
-def dfs_successors(G, source=None):
+def dfs_successors(G, source=None, depth_limit=None):
     """Return dictionary of successors in depth-first-search from source.
-
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     succ: dict
        A dictionary with nodes as keys and list of successor nodes as values.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -162,33 +162,31 @@ def dfs_successors(G, source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
     d = defaultdict(list)
-    for s,t in dfs_edges(G,source=source):
+    for s, t in dfs_edges(G, source=source, depth_limit=depth_limit):
         d[s].append(t)
     return dict(d)
 
 
-def dfs_postorder_nodes(G,source=None):
+def dfs_postorder_nodes(G, source=None, depth_limit=None):
     """Produce nodes in a depth-first-search post-ordering starting
     from source.
-
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     nodes: generator
        A generator of nodes in a depth-first-search post-ordering.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -200,34 +198,33 @@ def dfs_postorder_nodes(G,source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
-    post=(v for u,v,d in nx.dfs_labeled_edges(G,source=source)
-          if d['dir']=='reverse')
+    post = (v for u, v, d in nx.dfs_labeled_edges(G, source=source,
+    depth_limit=depth_limit)
+          if d['dir'] == 'reverse')
     # potential modification: chain source to end of post-ordering
     # return chain(post,[source])
     return post
 
 
-def dfs_preorder_nodes(G, source=None):
+def dfs_preorder_nodes(G, source=None, depth_limit=None):
     """Produce nodes in a depth-first-search pre-ordering starting
     from source.
-
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     nodes: generator
        A generator of nodes in a depth-first-search pre-ordering.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -239,34 +236,33 @@ def dfs_preorder_nodes(G, source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
-    pre=(v for u,v,d in nx.dfs_labeled_edges(G,source=source)
-         if d['dir']=='forward')
+    pre = (v for u, v, d in nx.dfs_labeled_edges(G, source=source,
+    depth_limit=depth_limit)
+         if d['dir'] == 'forward')
     # potential modification: chain source to beginning of pre-ordering
     # return chain([source],pre)
     return pre
 
 
-def dfs_labeled_edges(G, source=None):
+def dfs_labeled_edges(G, source=None, depth_limit=None):
     """Produce edges in a depth-first-search (DFS) labeled by type.
-
     Parameters
     ----------
     G : NetworkX graph
-
     source : node, optional
        Specify starting node for depth-first search and return edges in
        the component reachable from source.
-
+    depth_limit : length, optional
+       Specify the maximum search depth.
     Returns
     -------
     edges: generator
        A generator of edges in the depth-first-search labeled with 'forward',
        'nontree', and 'reverse'.
-
     Examples
     --------
     >>> G = nx.Graph()
@@ -277,7 +273,7 @@ def dfs_labeled_edges(G, source=None):
     -----
     Based on http://www.ics.uci.edu/~eppstein/PADS/DFS.py
     by D. Eppstein, July 2004.
-
+    https://en.wikipedia.org/wiki/Depth-limited_search
     If a source is not specified then a source is chosen arbitrarily and
     repeatedly until all components in the graph are searched.
     """
@@ -290,24 +286,29 @@ def dfs_labeled_edges(G, source=None):
         # produce edges for components with source
         nodes = [source]
     visited = set()
+    if depth_limit is None:
+       depth_limit = len(G)
     for start in nodes:
         if start in visited:
             continue
-        yield start,start,{'dir':'forward'}
+        yield start, start, {'dir': 'forward'}
         visited.add(start)
-        stack = [(start,iter(G[start]))]
+        stack = [(start, depth_limit, iter(G[start]))]
         while stack:
-            parent,children = stack[-1]
+            parent, depth_now, children = stack[-1]
+            if not depth_now >= 1:
+                stack.pop()
+                continue
             try:
                 child = next(children)
                 if child in visited:
-                    yield parent,child,{'dir':'nontree'}
+                    yield parent, child, {'dir': 'nontree'}
                 else:
-                    yield parent,child,{'dir':'forward'}
+                    yield parent, child, {'dir': 'forward'}
                     visited.add(child)
-                    stack.append((child,iter(G[child])))
+                    stack.append((child, depth_now-1, iter(G[child])))
             except StopIteration:
                 stack.pop()
                 if stack:
-                    yield stack[-1][0],parent,{'dir':'reverse'}
-        yield start,start,{'dir':'reverse'}
+                    yield stack[-1][0], parent, {'dir': 'reverse'}
+        yield start, start, {'dir': 'reverse'}
